backend/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
websockets
python-dotenv
email-validator
